third_party/blink/web_tests/virtual/document-transition-wide-gamut/document-transition/capture-callback-exception-expected.txt

CONSOLE ERROR: Uncaught (in promise) ReferenceError: foo is not defined
